memory problem fix on make_system
I noticed that the game increasingly consuming memory.

After investigating with heaptrack and looking at the C generated code,
I discovered that was something to do with the self.draw_order vector
variable on sprite painter runner.

I tried to fix this for hours, but then I just realized that the problem
was in make_system generic.

It was an actually simple and obvious mistake, it was creating a new
runner for each frame, and the problem was only visible on sprite
painter because it's the only system that was creating a new vector per
frame.

Another extra thing: y_vec was also responsible, but just destroying it
resolves the problem.