refactor on Collider component and Collision Detection system
using `cycles_without_sync`instead of `was_sync` gives more information
to collision detection system.

In short, when a collider's cycles_without_sync is 0, it can create an
intersection with an "Enter" state, and when is 1, it can create with a
"Stay" state.

There is also a `is_immovable` field, which also helps the collision
detection system.